Write errors to stdout instead of stderr for Octopus capture
Put this on stdout because we have problems capturing the output in
powershell (used by octopus).  If we use 2>&1 to redirect the output of
migrator from Poweshell, something happens to that session such that it
cannot run scripts anymore.  Putting this on stdout allows Octopus to
capture the output without a problem.